# docs/openapi.yaml
openapi: 3.1.0

info:
  title: API do Sistema de Orçamentos
  description: >
    API REST responsável pelas regras de negócio e pelo acesso aos dados
    do Sistema de Orçamentos.
  version: 0.1.0

servers:
  - url: http://localhost:3000
    description: Ambiente local de desenvolvimento

tags:
  - name: Status
    description: Verificação do funcionamento da API.
  - name: Usuário
    description: Informações do usuário autenticado.
  - name: Perfil profissional
    description: Dados profissionais ou empresariais do prestador autenticado.
  - name: Clientes
    description: Operações relacionadas aos clientes salvos pelo prestador.
  - name: Orçamentos
    description: Operações relacionadas à criação e ao acompanhamento de orçamentos.

paths:
  /health:
    get:
      tags:
        - Status
      summary: Verifica o funcionamento da API
      operationId: getHealth
      security: []
      responses:
        '200':
          description: API disponível.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/me:
    get:
      tags:
        - Usuário
      summary: Consulta o usuário autenticado
      description: >
        Retorna os dados básicos do prestador identificado pelo token
        de acesso emitido pelo Auth0.
      operationId: getAuthenticatedUser
      security:
        - Auth0Bearer: []
      responses:
        '200':
          description: Usuário autenticado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticatedUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/company:
    get:
      tags:
        - Perfil profissional
      summary: Consulta os dados profissionais
      description: >
        Retorna os dados profissionais ou empresariais pertencentes ao
        prestador autenticado.
      operationId: getCompany
      security:
        - Auth0Bearer: []
      responses:
        '200':
          description: Dados profissionais encontrados.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Perfil profissional
      summary: Salva os dados profissionais
      description: >
        Cria ou atualiza os dados profissionais ou empresariais pertencentes
        ao prestador autenticado.
      operationId: saveCompany
      security:
        - Auth0Bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanySaveRequest'
      responses:
        '200':
          description: Dados profissionais salvos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/clients:
    get:
      tags:
        - Clientes
      summary: Lista os clientes salvos
      description: >
        Retorna os clientes ativos salvos pelo prestador autenticado para
        reutilização na criação de orçamentos. Quando não houver clientes,
        retorna uma lista vazia.
      operationId: listClients
      security:
        - Auth0Bearer: []
      responses:
        '200':
          description: Lista de clientes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Clientes
      summary: Cadastra um cliente
      description: >
        Cadastra um cliente e o associa automaticamente ao prestador
        autenticado. O identificador do prestador não é recebido no corpo
        da requisição.
      operationId: createClient
      security:
        - Auth0Bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientCreateRequest'
      responses:
        '201':
          description: Cliente cadastrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/quotes:
    get:
      tags:
        - Orçamentos
      summary: Lista os orçamentos do prestador
      description: >
        Retorna somente os orçamentos pertencentes ao prestador autenticado,
        ordenados do mais recente para o mais antigo. Quando não houver
        orçamentos, retorna uma lista vazia.
      operationId: listQuotes
      security:
        - Auth0Bearer: []
      responses:
        '200':
          description: Lista de orçamentos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Orçamentos
      summary: Cria um orçamento em rascunho
      description: >
        Cria um orçamento editável associado a um cliente do prestador
        autenticado. Os dados de contato do cliente e o endereço do serviço
        são copiados para preservar o histórico do orçamento.
      operationId: createQuoteDraft
      security:
        - Auth0Bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteCreateRequest'
      responses:
        '201':
          description: Orçamento criado em rascunho.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/quotes/{quoteId}:
    put:
      tags:
        - Orçamentos
      summary: Atualiza um orçamento em rascunho
      description: >
        Atualiza os dados editáveis de um orçamento pertencente ao prestador
        autenticado. Somente orçamentos em rascunho podem ser alterados.
      operationId: updateQuoteDraft
      security:
        - Auth0Bearer: []
      parameters:
        - name: quoteId
          in: path
          required: true
          description: Identificador do orçamento.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteUpdateRequest'
      responses:
        '200':
          description: Orçamento atualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/quotes/{quoteId}/confirm:
    post:
      tags:
        - Orçamentos
      summary: Confirma um orçamento em rascunho
      description: >
        Confirma um orçamento pertencente ao prestador autenticado, altera
        seu status para enviado, bloqueia novas edições e gera o token usado
        no link público. Exige um perfil profissional ativo e copia seus dados
        para o orçamento, preservando o histórico.
      operationId: confirmQuote
      security:
        - Auth0Bearer: []
      parameters:
        - name: quoteId
          in: path
          required: true
          description: Identificador do orçamento.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Orçamento confirmado e preparado para compartilhamento.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/public/quotes/{publicToken}:
    get:
      tags:
        - Orçamentos
      summary: Consulta um orçamento pelo link público
      description: >
        Retorna os dados públicos de um orçamento confirmado. Não exige
        autenticação, pois o token presente no link funciona como chave de acesso.
      operationId: getPublicQuote
      security: []
      parameters:
        - name: publicToken
          in: path
          required: true
          description: Token público gerado na confirmação do orçamento.
          schema:
            type: string
            minLength: 64
            maxLength: 64
            pattern: '^[0-9a-f]{64}$'
      responses:
        '200':
          description: Dados públicos do orçamento.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicQuote'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/public/quotes/{publicToken}/respond:
    post:
      tags:
        - Orçamentos
      summary: Registra a resposta do cliente
      description: >
        Permite que o cliente aceite ou recuse um orçamento por meio do
        token público. Um orçamento só pode receber uma resposta.
      operationId: respondToPublicQuote
      security: []
      parameters:
        - name: publicToken
          in: path
          required: true
          description: Token público gerado na confirmação do orçamento.
          schema:
            type: string
            minLength: 64
            maxLength: 64
            pattern: '^[0-9a-f]{64}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteResponseRequest'
      responses:
        '200':
          description: Resposta registrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicQuote'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/quotes/{quoteId}/corrections:
    post:
      tags:
        - Orçamentos
      summary: Cria uma correção de um orçamento recusado
      description: >
        Cria um novo orçamento em rascunho copiando os dados de um orçamento
        recusado pertencente ao prestador autenticado. O orçamento original
        é preservado no histórico.
      operationId: createQuoteCorrection
      security:
        - Auth0Bearer: []
      parameters:
        - name: quoteId
          in: path
          required: true
          description: Identificador do orçamento recusado.
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Correção criada em rascunho.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    Auth0Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de acesso emitido pelo Auth0.

  responses:
    BadRequest:
      description: Dados enviados são inválidos.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Token ausente, inválido ou expirado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    Forbidden:
      description: Usuário autenticado, mas sem permissão para a operação.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    NotFound:
      description: Recurso não encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: O estado atual do recurso não permite esta operação.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Erro interno inesperado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    HealthResponse:
      type: object
      additionalProperties: false
      required:
        - status
      properties:
        status:
          type: string
          examples:
            - 'ok'

    AuthenticatedUser:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - email
      properties:
        id:
          type: string
          format: uuid
          description: Identificador interno do usuário no Sistema de Orçamentos.
          examples:
            - 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          description: Nome do usuário autenticado.
          examples:
            - Maria da Silva
        email:
          type: string
          format: email
          description: E-mail do usuário autenticado.
          examples:
            - maria@example.com

    CompanySaveRequest:
      type: object
      additionalProperties: false
      required:
        - name
        - email
        - phone
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 150
          description: Nome profissional ou nome da empresa.
        email:
          type: string
          format: email
          maxLength: 320
          description: E-mail comercial do prestador.
        phone:
          type: string
          minLength: 8
          maxLength: 30
          description: Telefone comercial do prestador.
        taxId:
          type: string
          maxLength: 20
          description: CPF ou CNPJ opcional, com 11 ou 14 dígitos.
        address:
          $ref: '#/components/schemas/Address'

    Company:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - email
        - phone
      properties:
        id:
          type: string
          format: uuid
          description: Identificador dos dados profissionais.
        name:
          type: string
          description: Nome profissional ou nome da empresa.
        email:
          type: string
          format: email
          description: E-mail comercial do prestador.
        phone:
          type: string
          description: Telefone comercial do prestador.
        taxId:
          type: string
          description: CPF ou CNPJ do prestador.
        address:
          $ref: '#/components/schemas/Address'

    ClientListResponse:
      type: object
      additionalProperties: false
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Client'

    ClientCreateRequest:
      type: object
      additionalProperties: false
      required:
        - name
        - email
        - phone
        - address
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 150
          description: Nome completo ou nome empresarial do cliente.
        email:
          type: string
          format: email
          maxLength: 320
          description: E-mail utilizado para o envio do orçamento.
        phone:
          type: string
          minLength: 8
          maxLength: 30
          description: Telefone para contato com o cliente.
        address:
          $ref: '#/components/schemas/Address'

    QuoteUpdateRequest:
      type: object
      additionalProperties: false
      required:
        - description
        - totalAmount
        - serviceDate
        - serviceAddress
      properties:
        description:
          type: string
          minLength: 1
          maxLength: 10000
          description: Descrição livre dos serviços e valores do orçamento.
        totalAmount:
          type: string
          pattern: '^\d{1,10}(\.\d{1,2})?$'
          description: Valor total informado manualmente, maior que zero.
          examples:
            - '1500.50'
        serviceDate:
          type: string
          format: date
          description: Data prevista para a execução do serviço.
          examples:
            - '2026-10-15'
        serviceAddress:
          $ref: '#/components/schemas/Address'
        locationNotes:
          type: string
          maxLength: 2000
          description: Observações opcionais sobre o local do serviço.

    Client:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - email
        - phone
        - address
      properties:
        id:
          type: string
          format: uuid
          description: Identificador do cliente.

        name:
          type: string
          description: Nome completo ou nome empresarial do cliente.
        email:
          type: string
          format: email
          description: E-mail utilizado para o envio do orçamento.
        phone:
          type: string
          description: Telefone para contato com o cliente.
        address:
          $ref: '#/components/schemas/Address'

    Address:
      type: object
      additionalProperties: false
      required:
        - street
        - number
        - postalCode
        - district
        - city
        - state
      properties:
        street:
          type: string
          description: Rua, avenida ou outra identificação do logradouro.
        number:
          type: string
          description: Número do endereço.
        complement:
          type: string
          maxLength: 150
          description: Complemento opcional do endereço.
        postalCode:
          type: string
          description: CEP.
        district:
          type: string
          description: Bairro.
        city:
          type: string
          description: Cidade.
        state:
          type: string
          description: Estado.

    QuoteCreateRequest:
      type: object
      additionalProperties: false
      required:
        - clientId
        - description
        - totalAmount
        - serviceDate
        - serviceAddress
      properties:
        clientId:
          type: string
          format: uuid
          description: Identificador do cliente cadastrado.
        description:
          type: string
          minLength: 1
          maxLength: 10000
          description: Descrição livre dos serviços e valores do orçamento.
        totalAmount:
          type: string
          pattern: '^\d{1,10}(\.\d{1,2})?$'
          description: Valor total informado manualmente, maior que zero.
          examples:
            - '1500.50'
        serviceDate:
          type: string
          format: date
          description: Data prevista para a execução do serviço.
          examples:
            - '2026-10-15'
        serviceAddress:
          $ref: '#/components/schemas/Address'
        locationNotes:
          type: string
          maxLength: 2000
          description: Observações opcionais sobre o local do serviço.

    QuoteClientSnapshot:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - email
        - phone
      properties:
        id:
          type: string
          format: uuid
          description: Identificador do cliente cadastrado.
        name:
          type: string
          description: Nome do cliente no momento da criação do orçamento.
        email:
          type: string
          format: email
          description: E-mail do cliente no momento da criação do orçamento.
        phone:
          type: string
          description: Telefone do cliente no momento da criação do orçamento.

    QuoteProviderSnapshot:
      type: object
      additionalProperties: false
      required:
        - name
        - email
        - phone
      properties:
        name:
          type: string
          description: Nome profissional no momento da confirmação do orçamento.
        email:
          type: string
          format: email
          description: E-mail comercial no momento da confirmação do orçamento.
        phone:
          type: string
          description: Telefone comercial no momento da confirmação do orçamento.
        taxId:
          type: string
          description: CPF ou CNPJ no momento da confirmação do orçamento.
        address:
          $ref: '#/components/schemas/Address'

    Quote:
      type: object
      additionalProperties: false
      required:
        - id
        - quoteNumber
        - client
        - description
        - totalAmount
        - serviceDate
        - serviceAddress
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Identificador do orçamento.
        quoteNumber:
          type: integer
          minimum: 1
          description: Número sequencial utilizado para identificar o orçamento.
          examples:
            - 1
        client:
          $ref: '#/components/schemas/QuoteClientSnapshot'
        description:
          type: string
          description: Descrição livre do orçamento.
        totalAmount:
          type: string
          description: Valor total preservado com duas casas decimais.
          examples:
            - '1500.50'
        serviceDate:
          type: string
          format: date
          description: Data prevista para a execução do serviço.
        serviceAddress:
          $ref: '#/components/schemas/Address'
        locationNotes:
          type: string
          description: Observações opcionais sobre o local do serviço.
        status:
          type: string
          enum:
            - DRAFT
            - SENT
            - ACCEPTED
            - REJECTED
          description: Situação atual do orçamento.
        publicToken:
          type: string
          minLength: 64
          maxLength: 64
          pattern: '^[0-9a-f]{64}$'
          description: Token utilizado para acessar o orçamento pelo link público.
        sentAt:
          type: string
          format: date-time
          description: Data e horário em que o orçamento foi confirmado.
        respondedAt:
          type: string
          format: date-time
          description: Data e horário da resposta do cliente.
        rejectionReason:
          type: string
          maxLength: 2000
          description: Motivo informado pelo cliente em caso de recusa.
        createdAt:
          type: string
          format: date-time
          description: Data e horário de criação do orçamento.
        correctedFromId:
          type: string
          format: uuid
          description: Identificador do orçamento recusado que originou esta correção.

    QuoteResponseRequest:
      type: object
      additionalProperties: false
      required:
        - decision
      properties:
        decision:
          type: string
          enum:
            - ACCEPTED
            - REJECTED
          description: Decisão tomada pelo cliente.
        reason:
          type: string
          maxLength: 2000
          description: Motivo opcional informado em caso de recusa.

    PublicQuote:
      type: object
      additionalProperties: false
      required:
        - quoteNumber
        - clientName
        - description
        - totalAmount
        - serviceDate
        - serviceAddress
        - status
        - sentAt
      properties:
        quoteNumber:
          type: integer
          minimum: 1
          description: Número sequencial utilizado para identificar o orçamento.
        clientName:
          type: string
          description: Nome do cliente no momento da criação do orçamento.
        provider:
          $ref: '#/components/schemas/QuoteProviderSnapshot'
        description:
          type: string
          description: Descrição livre do orçamento.
        totalAmount:
          type: string
          description: Valor total preservado com duas casas decimais.
          examples:
            - '1500.50'
        serviceDate:
          type: string
          format: date
          description: Data prevista para a execução do serviço.
        serviceAddress:
          $ref: '#/components/schemas/Address'
        locationNotes:
          type: string
          description: Observações opcionais sobre o local do serviço.
        status:
          type: string
          enum:
            - SENT
            - ACCEPTED
            - REJECTED
          description: Situação atual do orçamento.
        sentAt:
          type: string
          format: date-time
          description: Data e horário em que o orçamento foi confirmado.
        respondedAt:
          type: string
          format: date-time
          description: Data e horário da resposta do cliente.
        rejectionReason:
          type: string
          maxLength: 2000
          description: Motivo informado pelo cliente em caso de recusa.

    QuoteListResponse:
      type: object
      additionalProperties: false
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Quote'

    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Código estável que identifica o tipo do erro.
          examples:
            - 'UNAUTHORIZED'
        message:
          type: string
          description: Mensagem compreensível sobre o erro.
          examples:
            - 'É necessário apresentar um token de acesso válido.'
        details:
          type: array
          description: Detalhes adicionais, principalmente sobre validações.
          items:
            type: object
            additionalProperties: false
            properties:
              field:
                type: string
              message:
                type: string